refactor(vertex): split client modules for GH727
